Simply and re-organize the process of making field descriptors sealable.
This avoids the usage of metaclasses and remove a ton of boilerplate.
