+++ tests/test_models.py
@@ -8,7 +8,7 @@ class SealableModelTests(SimpleTestCase):
     def test_sealed_instance_deferred_attribute_access(self):
         instance = SeaLion.from_db('default', ['id'], [1])
         instance.seal()
-        message = "Cannot fetch deferred fields weight on a sealed object."
+        message = "Cannot fetch deferred field weight on a sealed object."
         with self.assertRaisesMessage(SealedObject, message):
             instance.weight
